react-dialog(refactor): remove useDialog
